`@remotion/studio`: Add freeze frame action to timeline layers
